Obj props logical eval
I made the live property
optional in case a project doesn't have a live version, and now instead of linking to a potencially non-existent
live preview, It'll fallback to the project's repo.

Also the icon won't show up if the live version doesn't exist.

I'll need to do the same eval for the project's repo -- when I need it.